Moved interfaces into contract folder
