+++ src/Typesystem/Contract/NamedTypeVisitor.php
@@ -2,7 +2,7 @@
 
 declare(strict_types = 1);
 
-namespace Graphpinator\Typesystem;
+namespace Graphpinator\Typesystem\Contract;
 
 interface NamedTypeVisitor
 {
