--- src/Typesystem/Type.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types = 1);
-
-namespace Graphpinator\Typesystem;
-
-interface Type
-{
-    public function accept(\Graphpinator\Typesystem\TypeVisitor $visitor) : mixed;
-}
